chore(plugins) move useBR, tabReplace, and HTML merging into internal plugins
