+++ test/fixtures/expect/customtabreplace.txt
